(WIP, LEVEL EDITOR DOESN'T WORK YET)
remove xxxSave properties of Block
and we use new array levelObjectsSave, etc. in Game to implement the save/load feature
add copy constructor to GameObjects and Block in order to implement the save/load feature
added ObservePointer to ScriptUserClass, etc. as a reinventing-the-wheel of std::weak_ptr
use ObservePointer instead of Block* to save objCurrentStand, etc. in Block, Player and Game
remove save/load/reset function in GameObjects and Block (but still in Scenery)
some other fixes